casting222

+++ src/android/Screenshot.java
@@ -97,11 +97,11 @@ public class Screenshot extends CordovaPlugin {
 
     private void saveScreenshot(Bitmap bitmap, String format, String fileName, Integer quality, JSONObject crop) {
         try {
-			float dpr=bitmap.getWidth()/crop.getInt("actual_width");
-			float width=crop.getInt("width") * dpr;
-			float height=crop.getInt("height") * dpr;
-			float top=crop.getInt("top") * dpr;
-			float left=crop.getInt("left") * dpr;
+			float dpr=(float) bitmap.getWidth()/(float) crop.getInt("actual_width");
+			float width=(float) crop.getInt("width") * dpr;
+			float height=(float) crop.getInt("height") * dpr;
+			float top=(float) crop.getInt("top") * dpr;
+			float left=(float) crop.getInt("left") * dpr;
 			Log.d("SCREENSHOT", "Bitmap Size ("+bitmap.getWidth()+"x"+bitmap.getHeight()+")");
 			Log.d("SCREENSHOT", "Bounds (top:"+top+" left:"+left+" width:"+width+" height:"+height+")");
 			Log.d("SCREENSHOT", "Actual Width:"+crop.getInt("actual_width")+ " Actual Height:"+crop.getInt("actual_height"));
